+++ Js/Database.js
@@ -53,6 +53,63 @@ function createUser(email,pass,name){
     });
 }
 
-module.exports = {getUserById, getUserByEmail,createUser};
+async function searchEvent(req,res){
+    let searchString = "SELECT * FROM EVENT";
+    let fileterOptions="";
+    for(let name in req.body)
+    {
+        if(name =="search")
+        {
+            fileterOptions+= " name LIKE '" + req.body[name] + "'";
+        }
+        
+        else
+        {
+            if(Array.isArray(req.body[name]))
+            {
+                fileterOptions+= name + " BETWEEN "+ req.body[name][0] + " AND " + req.body[name][1];
+            }
+            else
+            {
+                fileterOptions+= " "+name + " == '" + req.body[name] + "'";
+            }
+            
+        }
+        fileterOptions += " AND"
+    }
+    fileterOptions = fileterOptions.substring(0,fileterOptions.length-3)
+    if(fileterOptions!="")
+    {
+        searchString+= " WHERE" + fileterOptions;
+    }
+
+    console.log(searchString)
+    const result = await pool.query(searchString)
+    res.send(result)   
+}
+
+async function createEvent(req,res)
+{
+    let createString = "INSERT INTO event";
+    let variabels="(";
+    let values="(";
+    for(let name in req.body)
+    {
+        variabels+=name+",";
+        values+= "'"+req.body[name]+"'";
+        values+=",";
+    }
+    variabels = variabels.substring(0,variabels.length-2)+")";
+    variabels = variabels.substring(0,values.length-2)+")";
+
+    createString+=" "+ variabels + " VALUES "+ values;    
+
+    await pool.query(createString);
+}
+
+
+
+
+module.exports = {getUserById, getUserByEmail,createUser,searchEvent,createEvent};
